Extract IUndoRenameLogic narrow port; inject it into UndoManager in place of IRenameLogic (ADR-0005 Phase 3)
Mirrors #3355 (IUndoPluginNotifier). UndoManager injected the full, MEF/WinForms-entangled IRenameLogic but called only three of its ~12 methods (HandleRename, GetChangesForRenamedVariable, ApplyVariableRenameChanges), all with headless GumDataTypes/ToolsUtilities parameters.

This extracts those three calls into a narrow port, IUndoRenameLogic, in the headless Gum.Presentation assembly, and injects that port into UndoManager in place of IRenameLogic. The three methods' closure pulled five small data types out of the tool's Gum/Logic/RenameLogic.cs (NameChangeAction + SideOfEquals enums; VariableChange / VariableReferenceChange / VariableChangeResponse) which are relocated to Tools/Gum.Presentation/Logic/RenameChangeTypes.cs, keeping the Gum.Logic namespace so the ~20 tool-side consumers compile unchanged.

RenameLogic's signatures matched exactly, so ': IRenameLogic, IUndoRenameLogic' was all that was needed (no adapter). DI registers RenameLogic as its own singleton and resolves both interfaces to it, so rename calls fire as before. This is the last shell-only ctor dependency UndoManager held; with it ported, all of UndoManager's ctor deps are now headless, a prerequisite for relocating the class itself headless in a later PR. This PR does NOT move UndoManager.

Tests: the two existing rename-delegation pinning tests now mock Mock<IUndoRenameLogic>. Added PerformUndo_OnElementRename_ShouldDelegateToRenameLogicHandleRename to pin the third call (HandleRename) - red-first verified (disabling the seam produced 'No invocations performed'). Full GumToolUnitTests suite green: 1000 passed, 0 failed, 4 pre-existing skips. AllPluginsCompositionTests confirms plugins still compose via MEF. Gum.Presentation builds standalone (zero WPF/WinForms leak).
